// src/libs/generated/clients/population.ts
/**
 * Generated by Orval 🍺
 * Do not edit manually.
 * 都道府県別人口構成API
 * RESAS(地域経済分析システム)のデータを元にした、都道府県別の人口構成を取得するAPI。

 * OpenAPI spec version: 1.0.0
 * biome-ignore-all lint: due to automatically generated contents
 */
import type {
  Error,
  GetPopulationCompositionPerYearParams,
  PopulationCompositionPerYearResponse,
} from './api.schemas';

/**
 * 指定された都道府県の年齢別人口構成データを取得します
 * @summary 都道府県別人口構成取得
 */
export type getPopulationCompositionPerYearResponse200 = {
  data: PopulationCompositionPerYearResponse;
  status: 200;
};

export type getPopulationCompositionPerYearResponse400 = {
  data: Error;
  status: 400;
};

export type getPopulationCompositionPerYearResponse403 = {
  data: Error;
  status: 403;
};

export type getPopulationCompositionPerYearResponse404 = {
  data: Error;
  status: 404;
};

export type getPopulationCompositionPerYearResponse500 = {
  data: Error;
  status: 500;
};

export type getPopulationCompositionPerYearResponseSuccess =
  getPopulationCompositionPerYearResponse200 & {
    headers: Headers;
  };
export type getPopulationCompositionPerYearResponseError = (
  | getPopulationCompositionPerYearResponse400
  | getPopulationCompositionPerYearResponse403
  | getPopulationCompositionPerYearResponse404
  | getPopulationCompositionPerYearResponse500
) & {
  headers: Headers;
};

export type getPopulationCompositionPerYearResponse =
  | getPopulationCompositionPerYearResponseSuccess
  | getPopulationCompositionPerYearResponseError;

export const getGetPopulationCompositionPerYearUrl = (
  params: GetPopulationCompositionPerYearParams,
) => {
  const normalizedParams = new URLSearchParams();

  Object.entries(params || {}).forEach(([key, value]) => {
    if (value !== undefined) {
      normalizedParams.append(key, value === null ? 'null' : value.toString());
    }
  });

  const stringifiedParams = normalizedParams.toString();

  return stringifiedParams.length > 0
    ? `https://frontend-engineer-codecheck-api.mirai.yumemi.io/api/v1/population/composition/perYear?${stringifiedParams}`
    : `https://frontend-engineer-codecheck-api.mirai.yumemi.io/api/v1/population/composition/perYear`;
};

export const getPopulationCompositionPerYear = async (
  params: GetPopulationCompositionPerYearParams,
  options?: RequestInit,
): Promise<getPopulationCompositionPerYearResponse> => {
  const res = await fetch(getGetPopulationCompositionPerYearUrl(params), {
    ...options,
    method: 'GET',
  });

  const body = [204, 205, 304].includes(res.status) ? null : await res.text();

  const data: getPopulationCompositionPerYearResponse['data'] = body
    ? JSON.parse(body)
    : {};
  return {
    data,
    status: res.status,
    headers: res.headers,
  } as getPopulationCompositionPerYearResponse;
};